fix fs functions

+++ posix/posixtypes.lua
@@ -1,6 +1,6 @@
 require("ffi").cdef[[
-    typedef unsigned int ino_t;
-    typedef int off_t;
+    typedef size_t ino_t;
+    typedef size_t off_t;
     typedef int pid_t;
     typedef void *locale_t;
 ]]
